requirements.txt

bitsandbytes 
accelerate 
timm
tqdm 
transformers
ftfy 
regex
git+https://github.com/openai/CLIP.git
torch
torchvision
pillow
numpy
